Changed edit popup to modal to be responsive
